users.js 주석 추가, /overlap/phone 로직 수정, error code를 config/code.js로 만듬
